Create generateAvatarColor File
